+++ classes/core/base_decorator.rb
