+++ classes/core/book.rb
