+++ classes/core/capitalize_decorator.rb
