+++ classes/core/classroom.rb
