+++ classes/core/corrector.rb
