+++ classes/core/nameable.rb
