+++ classes/core/person.rb
