+++ classes/core/rental.rb
