+++ classes/core/student.rb
