+++ classes/core/trimmer_decorator.rb
