+++ classes/menu_options/exit.rb
