+++ classes/menu_options/list_rentals.rb
