+++ classes/storageIO/storage_manager.rb
