+++ data/books.json
@@ -1,1 +1,1 @@
-[{"title":"TitleI","author":"Author1"},{"title":"TitleII","author":"Author2"},{"title":"TitleIII","author":"Author3"},{"title":"TitleIV","author":"Author4"},{"title":"TitleV","author":"Author5"},{"title":"The lord of the rings","author":"Tolkien"}]
+[{"title":"TitleI","author":"Author1"},{"title":"TitleII","author":"Author2"},{"title":"TitleIII","author":"Author3"},{"title":"TitleIV","author":"Author4"},{"title":"TitleV","author":"Author5"},{"title":"The lord of the rings","author":"Tolkien"},{"title":"Clean Code","author":"Robert C. Martin"}]
