+++ data/people.json
@@ -1,1 +1,1 @@
-{"students":[{"name":"Name1","age":12,"permission":null},{"name":"Name2","age":122,"permission":null},{"name":"Name3","age":123,"permission":null},{"name":"Seba","age":23,"permission":false}],"teachers":[{"name":"Spec1","age":"Name1","specialization":27},{"name":"Spec2","age":"Name2","specialization":68},{"name":"Spec3","age":"Name3","specialization":48}]}
+{"students":[{"name":"Name1","age":12,"permission":null},{"name":"Name2","age":122,"permission":null},{"name":"Name3","age":123,"permission":null},{"name":"Seba","age":23,"permission":null},{"name":"Jesus","age":26,"permission":false}],"teachers":[{"name":27,"age":"Spec1","specialization":"Name1"},{"name":68,"age":"Spec2","specialization":"Name2"},{"name":48,"age":"Spec3","specialization":"Name3"}]}
