+++ data/rentals.json
@@ -1,1 +1,1 @@
-[{"date":"22/22/2022","book":"TitleV","person_name":"Name3"},{"date":"20/22/2022","book":"The lord of the rings","person_name":"Seba"}]
+[{"date":"22/22/2022","book":"TitleV","person_name":"Name3"},{"date":"20/22/2022","book":"The lord of the rings","person_name":"Seba"},{"date":"2022/07/26","book":"The lord of the rings","person_name":"Jesus"},{"date":"2022/07/26","book":"Clean Code","person_name":"Jesus"}]
